Make money handling decimal-exact and transfers concurrency-safe
Store wallet balances as NUMERIC(15,2) instead of Float so money is exact.
Add utils/money.py with to_decimal (coerce amounts before arithmetic, since
Decimal + float raises TypeError) and to_amount (cast balances to float for
JSON, since Flask serializes raw Decimal as a string). Apply both across all
wallet arithmetic and balance responses.

Rewrite the internal transfer (services/transfer_service.py) to be
concurrency-safe:
- fix the ledger insert, which used columns that don't exist on Transaction
  (sender_id/receiver_id/reference) and crashed every call
- lock both wallet rows FOR UPDATE before reading balances, in a deterministic
  (lowest user_id first) order to avoid A<->B deadlocks
- re-check balance and wallet status under the lock
- reject self-transfers

+++ Backend/models/wallet.py
@@ -18,9 +18,9 @@ class Wallet(db.Model):
     )
 
     balance = db.Column(
-        db.Float,
+        db.Numeric(15, 2),
         nullable=False,
-        default=0.0
+        default=0
     )
 
     currency = db.Column(
